Fix behavior around pause and the progress bar
This means calling `gauge.disable()` when pausing if progressEnabled to
ensure that the progress bar is cleared and won't be written to the screen
while the log is disabled.

And similarly calling `gauge.enable()` when resuming if progressEnabled.

And finally patching `log.enableProgress()` to ONLY skip calling
`gauge.enable()` when called while paused.