fix(refactor): update deps, lint, test coverage
Also fixes a bug in enableProgress that was enabling the gauge even if
the log was paused.

Coverage for branches is downgraded to 95 so that CI is green. This is a
compromise to get us up to date in dependencies.  The coverage is MORE
than it was before, and we can keep making it better but I don't think
it should block our updating of dependencies.